+++ components/SigngleArticle.tsx
