--- pages/personalpublic.tsx
@@ -1,269 +1,0 @@
-import Head from "next/head";
-import React from "react";
-
-import BlogPostCard21 from "../components/blog-post-card21";
-
-const Personalpublic = () => {
-  return (
-    <>
-      <div className="personalpublic-container">
-        <Head>
-          <title>personalpublic - Flower</title>
-          <meta property="og:title" content="personalpublic - Flower" />
-        </Head>
-        <section className="personalpublic-description">
-          <div className="personalpublic-container04">
-            <div className="personalpublic-container05">
-              <div className="personalpublic-container06">
-                <img
-                  alt="image"
-                  src="https://ih1.redbubble.net/image.3114059041.7813/st,small,845x845-pad,1000x1000,f8f8f8.jpg"
-                  loading="lazy"
-                  className="personalpublic-image02"
-                />
-              </div>
-            </div>
-            <div className="personalpublic-container07">
-              <div className="personalpublic-container08">
-                <div className="personalpublic-container09">
-                  <span className="personalpublic-text12">BT21_RJ012525</span>
-                  <img alt="image" src="/playground_assets/flower11-200h.png" className="personalpublic-image03" />
-                </div>
-                <div className="personalpublic-container10">
-                  <img
-                    alt="image"
-                    src="/playground_assets/pastedimage-95s-200h.png"
-                    className="personalpublic-image04"
-                  />
-                  <span className="personalpublic-text13">0x4202043D9ff98a4e8C64b075dBF4Cb3eE5EfF528</span>
-                </div>
-              </div>
-              <div className="personalpublic-container11">
-                <span>
-                  <span> RJ喜歡做菜也喜歡享受美食。</span>
-                  <br></br>
-                  <span>用蓬鬆的毛髮和一顆溫暖的心撫慰每個人。</span>
-                </span>
-              </div>
-            </div>
-          </div>
-        </section>
-        <section className="personalpublic-collection">
-          <div className="personalpublic-container12">
-            <button className="personalpublic-button1 button">個人文章</button>
-            <button type="button" className="personalpublic-button2 button">
-              收藏花
-            </button>
-          </div>
-          <div className="personalpublic-container13">
-            <div className="personalpublic-container14">
-              <div className="personalpublic-blog">
-                <div className="personalpublic-container15">
-                  <BlogPostCard21
-                    profile_src="https://images.unsplash.com/photo-1611232658409-0d98127f237f?ixid=Mnw5MTMyMXwwfDF8c2VhcmNofDIzfHxwb3J0cmFpdCUyMHdvbWFufGVufDB8fHx8MTYyNjQ1MDU4MQ&amp;ixlib=rb-1.2.1&amp;h=1200"
-                    rootClassName="rootClassName7"
-                  ></BlogPostCard21>
-                </div>
-                <div className="personalpublic-container16">
-                  <BlogPostCard21
-                    image_src="https://images.unsplash.com/photo-1595565312451-23051ab0666c?ixid=Mnw5MTMyMXwwfDF8c2VhcmNofDI0fHxwb3J0cmFpdCUyMHdvbWFufGVufDB8fHx8MTYyNjQ1MDU4MQ&amp;ixlib=rb-1.2.1&amp;h=1000"
-                    profile_src="https://images.unsplash.com/photo-1438761681033-6461ffad8d80?ixid=Mnw5MTMyMXwwfDF8c2VhcmNofDExfHxwb3J0cmFpdCUyMHdvbWFufGVufDB8fHx8MTYyNjQ1MDU4MQ&amp;ixlib=rb-1.2.1&amp;h=1200"
-                    rootClassName="rootClassName8"
-                  ></BlogPostCard21>
-                </div>
-                <BlogPostCard21
-                  image_src="https://images.unsplash.com/photo-1519699047748-de8e457a634e?ixid=Mnw5MTMyMXwwfDF8c2VhcmNofDIxfHxwb3J0cmFpdCUyMHdvbWFuJTIwY29sb3J8ZW58MHx8fHwxNjI2NDUwNTU4&amp;ixlib=rb-1.2.1&amp;h=1000"
-                  profile_src="https://images.unsplash.com/photo-1531123897727-8f129e1688ce?ixid=Mnw5MTMyMXwwfDF8c2VhcmNofDE0fHxwb3J0cmFpdCUyMHdvbWFufGVufDB8fHx8MTYyNjQ1MDU4MQ&amp;ixlib=rb-1.2.1&amp;h=1200"
-                  rootClassName="rootClassName6"
-                ></BlogPostCard21>
-              </div>
-            </div>
-            <div className="personalpublic-container17">
-              <div className="personalpublic-container18">
-                <h1>推薦最佳使用者</h1>
-                <ul className="list">
-                  <li className="list-item">
-                    <div className="personalpublic-container19">
-                      <img
-                        alt="image"
-                        src="https://ih1.redbubble.net/image.3114059041.7813/st,small,845x845-pad,1000x1000,f8f8f8.jpg"
-                        loading="lazy"
-                        className="personalpublic-image05"
-                      />
-                      <span>BT21_RJ012525</span>
-                      <img alt="image" src="/playground_assets/flower11-200h.png" className="personalpublic-image06" />
-                    </div>
-                  </li>
-                  <li className="list-item">
-                    <div className="personalpublic-container20">
-                      <img
-                        alt="image"
-                        src="https://pbs.twimg.com/ext_tw_video_thumb/1148064582071545858/pu/img/-asywjIzDbmCV3WS.jpg:large"
-                        loading="lazy"
-                        className="personalpublic-image07"
-                      />
-                      <span>BT21_CHIMMY2569825</span>
-                      <img alt="image" src="/playground_assets/flower2-200h.png" className="personalpublic-image08" />
-                    </div>
-                  </li>
-                  <li className="list-item">
-                    <div className="personalpublic-container21">
-                      <img
-                        alt="image"
-                        src="https://img-shoplineapp-com.s3.amazonaws.com/media/image_clips/5bd68b58627c770015f73b73/large.jpg?1540787029"
-                        loading="lazy"
-                        className="personalpublic-image09"
-                      />
-                      <span>BT21_KOYA01125625</span>
-                      <img alt="image" src="/playground_assets/flower3-200h.png" className="personalpublic-image10" />
-                    </div>
-                  </li>
-                </ul>
-              </div>
-              <div className="personalpublic-container22">
-                <h1>今日一花</h1>
-                <img alt="image" src="/playground_assets/flower11-200h.png" className="personalpublic-image11" />
-                <h1>經典之花：紅玫瑰</h1>
-                <span className="personalpublic-text24">
-                  <span>以花為貌、以鳥為聲</span>
-                  <br></br>
-                  <br></br>
-                  <span>喜歡玫瑰花的人浪漫、熱情、性格開朗，對生活積極主動。</span>
-                </span>
-              </div>
-            </div>
-          </div>
-        </section>
-        <footer className="personalpublic-footer">
-          <div className="personalpublic-main">
-            <div className="personalpublic-branding">
-              <div className="personalpublic-heading">
-                <h2 className="personalpublic-logo1">Character</h2>
-                <p className="personalpublic-caption">
-                  Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et
-                  dolore.
-                </p>
-              </div>
-              <div className="personalpublic-socials">
-                <a
-                  href="https://twitter.com"
-                  target="_blank"
-                  rel="noreferrer noopener"
-                  className="personalpublic-twitter social button"
-                >
-                  <img alt="image" src="/playground_assets/twitter.svg" className="personalpublic-image12" />
-                </a>
-                <a
-                  href="https://discord.com"
-                  target="_blank"
-                  rel="noreferrer noopener"
-                  className="personalpublic-discord social button"
-                >
-                  <img alt="image" src="/playground_assets/discord.svg" className="personalpublic-image13" />
-                </a>
-              </div>
-            </div>
-            <div className="personalpublic-links">
-              <div className="personalpublic-list">
-                <h3 className="personalpublic-heading1">Site</h3>
-                <div className="personalpublic-items">
-                  <button className="personalpublic-link button-clean button">About</button>
-                  <button className="personalpublic-link1 button-clean button">Collection</button>
-                  <button className="personalpublic-link2 button-clean button">Roadmap</button>
-                  <button className="personalpublic-link3 button-clean button">Features</button>
-                </div>
-              </div>
-              <div className="personalpublic-list1">
-                <h3 className="personalpublic-heading2">Company</h3>
-                <div className="personalpublic-items1">
-                  <button className="personalpublic-link4 button-clean button">Team</button>
-                  <button className="personalpublic-link5 button-clean button">Press</button>
-                  <button className="personalpublic-link6 button-clean button">Terms</button>
-                  <button className="personalpublic-link7 button-clean button">Limitations</button>
-                  <button className="personalpublic-link8 button-clean button">Licenses</button>
-                </div>
-              </div>
-            </div>
-            <div className="personalpublic-socials1">
-              <a
-                href="https://twitter.com"
-                target="_blank"
-                rel="noreferrer noopener"
-                className="personalpublic-twitter1 social button"
-              >
-                <img alt="image" src="/playground_assets/twitter.svg" className="personalpublic-image14" />
-              </a>
-              <a
-                href="https://discord.com"
-                target="_blank"
-                rel="noreferrer noopener"
-                className="personalpublic-discord1 social button"
-              >
-                <img alt="image" src="/playground_assets/discord.svg" className="personalpublic-image15" />
-              </a>
-            </div>
-          </div>
-          <span className="personalpublic-copyright">© 2022 Character. All Rights Reserved.</span>
-        </footer>
-        <div>
-          {/* <DangerousHTML
-            html={`<script>
- /*
-  Accordion - Code Embed
-
-  const accordionContainers = document.querySelectorAll("[data-role="accordion-container"]"); // All accordion containers
-  const accordionContents = document.querySelectorAll("[data-role="accordion-content"]"); // All accordion content
-  const accordionIconsClosed = document.querySelectorAll("[data-role="accordion-icon-closed"]"); // All accordion closed icons
-  const accordionIconsOpen = document.querySelectorAll("[data-role="accordion-icon-open"]"); // All accordion open icons
-
-  accordionContents.forEach((accordionContent) => {
-      accordionContent.style.display = "none"; //Hides all accordion contents
-  });
-
-  accordionIconsClosed.forEach((icon) => {
-    icon.style.display = "flex"
-  })
-
-  accordionIconsOpen.forEach((icon) => {
-    icon.style.display = "none"
-  })
-
-  accordionContainers.forEach((accordionContainer, index) => {
-      accordionContainer.addEventListener("click", () => {
-          if (accordionContents[index].style.display === "flex") {
-              // If the accordion is already open, close it
-              accordionContents[index].style.display = "none";
-              accordionIconsClosed[index].style.display = "flex";
-              accordionIconsOpen[index].style.display = "none"
-          } else {
-              // If the accordion is closed, open it
-              accordionContents.forEach((accordionContent) => {
-                  accordionContent.style.display = "none"; //Hides all accordion contents
-              });
-
-              accordionIconsClosed.forEach((accordionIcon) => {
-                  accordionIcon.style.display = "flex"; // Resets all icon transforms to 0deg (default)
-              });
-
-              accordionIconsOpen.forEach((accordionIcon) => {
-                accordionIcon.style.display = "none";
-              })
-              
-              accordionContents[index].style.display = "flex"; // Shows accordion content
-              accordionIconsClosed[index].style.display = "none"; // Rotates accordion icon 180deg
-              accordionIconsOpen[index].style.display = "flex";
-          }
-      });
-  });
-</script>
-`}
-          ></DangerousHTML> */}
-        </div>
-        <div className="personalpublic-container23"></div>
-      </div>
-      <style jsx>{``}</style>
-    </>
-  );
-};
-
-export default Personalpublic;
